% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwebthemes-package.R
\docType{package}
\name{ggwebthemes-package}
\alias{ggwebthemes}
\alias{ggwebthemes-package}
\title{ggwebthemes: Some ggplot2 Themes Optimized for Blogs}
\description{
Some ggplot2 themes optimized for blogs and similar online usage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dataenthusiast.ca/}
  \item Report bugs at \url{https://gitlab.com/peterbar/ggwebthemes/-/issues}
}

}
\keyword{internal}
