% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwebthemes.R
\docType{package}
\name{ggwebthemes}
\alias{ggwebthemes}
\alias{ggwebthemes-package}
\title{ggwebthemes: Some ggplot2 themes optimized for blogs.}
\description{
This package provides a collection of some ggplot2 themes
optimized for use in blogs (and similar online usage).
}
\seealso{
\url{https://ggplot2.tidyverse.org/reference/ggtheme.html}
}
\author{
Petr Baranovskiy \email{peterbar@protonmail.com} \url{https://dataenthusiast.ca/}
}
