% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggwebthemes.R
\docType{package}
\name{ggwebthemes}
\alias{ggwebthemes}
\alias{ggwebthemes-package}
\title{Some ggplot2 themes optimized for blogs.}
\description{
Some ggplot2 themes optimized for blogs and similar online usage.
}
\seealso{
\url{https://ggplot2.tidyverse.org/reference/ggtheme.html}
}
\author{
Petr Baranovskiy \email{peterbar@protonmail.com} \url{https://dataenthusiast.ca/}
}
